// fe/src/lib/api/generated/model/resourceCategoryValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ResourceCategoryValue = typeof ResourceCategoryValue[keyof typeof ResourceCategoryValue];


export const ResourceCategoryValue = {
  GENERAL: 'GENERAL',
  COUNSELING: 'COUNSELING',
  MENTAL_HEALTH: 'MENTAL_HEALTH',
  ACADEMIC_SUPPORT: 'ACADEMIC_SUPPORT',
  CAREER: 'CAREER',
  WELLNESS: 'WELLNESS',
  FORMS_AND_GUIDES: 'FORMS_AND_GUIDES',
  OTHER: 'OTHER',
} as const;