// fe/src/lib/api/generated/model/resourceKindValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type ResourceKindValue = typeof ResourceKindValue[keyof typeof ResourceKindValue];


export const ResourceKindValue = {
  ARTICLE: 'ARTICLE',
  EXTERNAL_LINK: 'EXTERNAL_LINK',
  FILE: 'FILE',
} as const;